fix #1 - add timezone support
The command /set_utc_offset is added. How it works: if the user has set their time zone before entering data, then start_time will be converted to UTC before adding it to the database. If the user has not set the time zone, the start_time will be updated after it is set, by default the user's time is considered UTC.